ISSUE #4608 - extract selectionToolbar
